chore(refactor): Make separate blueprint for events and orders (#6615)
